refactored Settings
changed package name
fixed issues
