+++ src/main/java/sqltoregex/settings/RegExGenerator.java
@@ -1,4 +1,4 @@
-package sqltoregex.property;
+package sqltoregex.settings;
 
 import java.util.Set;
