+++ src/test/java/sqltoregex/deparser/ExpressionDeParserForRegExTest.java
@@ -6,7 +6,7 @@ import net.sf.jsqlparser.parser.CCJSqlParserUtil;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.xml.sax.SAXException;
-import sqltoregex.property.SettingsManager;
+import sqltoregex.settings.SettingsManager;
 
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.xpath.XPathExpressionException;
