+++ src/test/java/sqltoregex/settings/regexgenerator/SpellingMistakeTest.java
@@ -1,8 +1,8 @@
-package sqltoregex.property.regexgenerator;
+package sqltoregex.settings.regexgenerator;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import sqltoregex.property.SettingsOption;
+import sqltoregex.settings.SettingsOption;
 
 class SpellingMistakeTest {
     public SpellingMistake spellingMistake = new SpellingMistake(SettingsOption.DEFAULT);
